tesstrain/requirements.txt
numpy
opencv-python
Pillow
python-bidi
matplotlib
pandas
